Extract MultipleCIVCreateForm
